+++ CHANGES.md
@@ -13,6 +13,8 @@ Version 0.5.0
 
 To be released.
 
+ -  Fedify is now available on npm: [@fedify/fedify].  [[#24]]
+
  -  Abstract key-value store for caching.
 
      -  Added `KvStore` interface.
@@ -42,6 +44,8 @@ To be released.
      -  Added `exportSpki()` function.
      -  Added `importSpki()` function.
 
+[@fedify/fedify]: https://www.npmjs.com/package/@fedify/fedify
+[#24]: https://github.com/dahlia/fedify/discussions/24
 
 Version 0.4.0
 -------------